removes old "mathLib" and refactors mathlib2 to mathlib
